requirements.txt
# Pinned for reproducibility (matching current environment where applicable)
mcp==1.14.0
acp-sdk==0.0.6
a2a-sdk==0.3.5
agent-connect==0.3.6

anyio>=4.5
httpx>=0.28.1
httpx-sse>=0.4.0
uvicorn>=0.31.1
fastapi>=0.115.2
starlette>=0.27
sse-starlette>=2.1.3
numpy>=2.2.6
scipy>=1.14.1

